Separate Configuration class and refactor its use
